Refactorizado InputResumen para que reciba la lista de oradores y si debe o no estar visible.
