+++ frontend/src/minuta/InputResumen.js
@@ -1,37 +1,25 @@
 import Collapse from '@material-ui/core/Collapse';
 import { Card, CardContent, TextField } from '@material-ui/core';
-import React, { useState } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faChevronDown } from '@fortawesome/free-solid-svg-icons/faChevronDown';
-import { BotonParaAbrirResumen } from './Minuta.styled';
+import React from 'react';
 
-const InputResumen = ({ tema }) => {
-  const [isRecapVisible, setIsRecapCollapsed] = useState(false);
+const InputResumen = ({ oradores, isRecapVisible }) => {
 
-  const selected = () => (tema.oradores.actual
-    ? tema.oradores.actual.usuario.nombre
-    : tema.oradores.pasados[tema.oradores.pasados.length - 1].usuario.nombre);
+  const selected = () => (oradores.actual
+    ? oradores.actual.usuario.nombre
+    : oradores.pasados[oradores.pasados.length - 1].usuario.nombre);
 
-  const shouldBeVisible = () => tema.oradores.actual || tema.oradores.pasados.length > 0;
-
-  const buttonText = () => (isRecapVisible ? 'CERRAR EDICION' : 'ABRIR EDICION');
+  const shouldBeDisabled = () => !(oradores.actual || oradores.pasados.length > 0);
 
   return (
     <>
-      <BotonParaAbrirResumen
-        variant="outlined"
-        endIcon={<FontAwesomeIcon icon={faChevronDown}/>}
-        onClick={() => setIsRecapCollapsed(!isRecapVisible)}
-      >
-        {buttonText()}
-      </BotonParaAbrirResumen>
+
       <div>
         <Collapse in={isRecapVisible}>
           <Card>
             <CardContent>
-              {shouldBeVisible() ? <p>Estas editando el resumen de: {selected()}</p>
+              {!shouldBeDisabled() ? <p>Estas editando el resumen de: {selected()}</p>
                 : <p>Nadie habló</p>}
-              <TextField label="Resumen" variant="outlined" disabled={!shouldBeVisible()}/>
+              <TextField label="Resumen" variant="outlined" disabled={shouldBeDisabled()}/>
             </CardContent>
           </Card>
         </Collapse>
